+++ walkability/components/shade/utility/__init__.py
@@ -1,0 +1,7 @@
+import logging
+
+logging.getLogger('botocore').setLevel(logging.WARNING)
+logging.getLogger('boto3').setLevel(logging.WARNING)
+logging.getLogger('s3transfer').setLevel(logging.WARNING)
+logging.getLogger('urllib3').setLevel(logging.WARNING)
+logging.getLogger('rasterio').setLevel(logging.WARNING)